refactor: repository structure
