+++ backend/cmd/config/database.go
@@ -32,15 +32,40 @@ func createRegistrationsTable(db *sql.DB) error {
 		id INTEGER PRIMARY KEY AUTOINCREMENT,
 		name TEXT NOT NULL,
 		nickname TEXT NOT NULL UNIQUE,
+		email TEXT NOT NULL,
+		region TEXT NOT NULL,
 		project_idea TEXT NOT NULL,
-		teammate TEXT,
-		role TEXT NOT NULL CHECK(role IN ('frontend', 'designer', 'backend', 'fullstack')),
+		team_preference BOOLEAN NOT NULL,
+		desired_teammate TEXT,
 		created_at DATETIME DEFAULT CURRENT_TIMESTAMP
 	);
 	
+	CREATE INDEX IF NOT EXISTS idx_email ON registrations(email);
 	CREATE INDEX IF NOT EXISTS idx_nickname ON registrations(nickname);
-	CREATE INDEX IF NOT EXISTS idx_role ON registrations(role);
+	CREATE INDEX IF NOT EXISTS idx_region ON registrations(region);
 	CREATE INDEX IF NOT EXISTS idx_created_at ON registrations(created_at);
+
+	CREATE TABLE IF NOT EXISTS roles (
+		id INTEGER PRIMARY KEY AUTOINCREMENT,
+		name TEXT NOT NULL UNIQUE
+	);
+
+	CREATE TABLE IF NOT EXISTS technologies (
+		id INTEGER PRIMARY KEY AUTOINCREMENT,
+		name TEXT NOT NULL UNIQUE
+	);
+
+	CREATE TABLE IF NOT EXISTS registration_roles (
+		registration_id INTEGER NOT NULL,
+		role_id INTEGER NOT NULL,
+		PRIMARY KEY (registration_id, role_id)
+	);
+
+	CREATE TABLE IF NOT EXISTS registration_technologies (
+		registration_id INTEGER NOT NULL,
+		technology_id INTEGER NOT NULL,
+		PRIMARY KEY (registration_id, technology_id)
+	);
 	`
 
 	_, err := db.Exec(query)
