+++ backend/internal/registration/application/ports/registration.go
