+++ backend/internal/registration/application/ports/role.go
@@ -1,0 +1,10 @@
+package ports
+
+import (
+	"context"
+	"hackathon-pvc-backend/internal/registration/domain"
+)
+
+type RoleRepositoryPort interface {
+	Save(ctx context.Context, role *domain.Role) (*domain.Role, error)
+}
