+++ backend/internal/registration/infrastructure/rest/registration.go
@@ -1,1 +1,77 @@
 package rest
+
+import (
+	"encoding/json"
+	"net/http"
+
+	"hackathon-pvc-backend/internal/registration/application/services"
+	"hackathon-pvc-backend/internal/registration/domain"
+)
+
+type RegistrationHandler struct {
+	service *services.RegistrationService
+}
+
+func NewRegistrationHandler(service *services.RegistrationService) *RegistrationHandler {
+	return &RegistrationHandler{
+		service: service,
+	}
+}
+
+func (h *RegistrationHandler) CreateRegistration(w http.ResponseWriter, r *http.Request) {
+	var req CreateRegistrationRequest
+
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		h.writeErrorResponse(w, http.StatusBadRequest, "invalid JSON", err.Error())
+		return
+	}
+
+	registration, err := h.service.RegisterParticipant(
+		r.Context(),
+		req.Name,
+		req.Nickname,
+		req.Email,
+		req.Region,
+		req.ProjectIdea,
+		req.TeamPreference,
+		req.DesiredTeammate,
+	)
+	if err != nil {
+		h.writeErrorResponse(w, http.StatusBadRequest, "registration failed", err.Error())
+		return
+	}
+
+	// Convert to response DTO
+	response := h.toRegistrationResponse(registration)
+
+	// Write success response
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusCreated)
+	json.NewEncoder(w).Encode(map[string]interface{}{
+		"success": true,
+		"data":    response,
+		"message": "Registration created successfully",
+	})
+}
+
+func (h *RegistrationHandler) toRegistrationResponse(reg *domain.Registration) RegistrationResponse {
+	return RegistrationResponse{
+		ID:              reg.ID(),
+		Name:            reg.Name(),
+		Nickname:        reg.Nickname(),
+		Email:           reg.Email(),
+		Region:          reg.Region(),
+		ProjectIdea:     reg.ProjectIdea(),
+		DesiredTeammate: reg.DesiredTeammate(),
+		CreatedAt:       reg.CreatedAt().Format("2006-01-02T15:04:05Z07:00"),
+	}
+}
+
+func (h *RegistrationHandler) writeErrorResponse(w http.ResponseWriter, statusCode int, error, message string) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(statusCode)
+	json.NewEncoder(w).Encode(ErrorResponse{
+		Error:   error,
+		Message: message,
+	})
+}
